Solucionado errores de compilación(no todos)

+++ DeathTokenBase/DeathToken/HUD.cpp
@@ -44,12 +44,12 @@ HUD::HUD(GameState* gS) : GameObject(gS)
 {
 	gS->addObjectsUI(this);
 
-	balanceDescText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX(1600), relativeY(70),
-							   relativeX(50), relativeX(3), Text::DERECHA);
+	balanceDescText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX((float)1600.0f), relativeY((float)70.0f),
+							   relativeX((float)50.0f), relativeX((float)3.0f), Text::DERECHA);
 	balanceDescText->setMessage("SALDO");
 	gS->addObjectsUI(balanceDescText);
-	balanceText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX(1870), relativeY(70 + getNumberY(balance)),
-						   relativeX(getNumberSize(balance)), relativeX(3), Text::DERECHA);
+	balanceText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX((float)1870.0f), relativeY((float)70.0f + getNumberY(balance)),
+						   relativeX((float)getNumberSize(balance)), relativeX((float)3.0f), Text::DERECHA);
 	balanceText->setMessage(std::to_string(balance));
 	gS->addObjectsUI(balanceText);
 }
@@ -57,19 +57,19 @@ void
 HUD::refresh()
 {
 	balanceText->setMessage(std::to_string(balance));
-	balanceText->setPos(relativeX(1870), relativeY(70 + getNumberY(balance)));
-	balanceText->setSize(getNumberSize(balance));
+	balanceText->setPos(relativeX((float)1870), relativeY((float)70 + getNumberY(balance)));
+	balanceText->setSize((int)getNumberSize(balance));
 	balanceText->render();
 }
 
 HUDLobby::HUDLobby(GameState* gS) : HUD(gS)
 {
-	redSoulsDescText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX(1600), relativeY(150),
-								relativeX(50), relativeX(3), Text::DERECHA);
+	redSoulsDescText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX((float)1600), relativeY((float)150),
+								relativeX((float)50), relativeX((float)3), Text::DERECHA);
 	redSoulsDescText->setMessage("ALMAS");
 	gS->addObjectsUI(redSoulsDescText);
-	redSoulsText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX(1870), relativeY(150 + getNumberY(redSouls)),
-							relativeX(getNumberSize(redSouls)), relativeX(3), Text::DERECHA);
+	redSoulsText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX((float)1870), relativeY((float)150 + getNumberY(redSouls)),
+							relativeX((float)getNumberSize(redSouls)), relativeX((float)3), Text::DERECHA);
 	redSoulsText->setMessage(std::to_string(redSouls));
 	gS->addObjectsUI(redSoulsText);
 
@@ -80,19 +80,19 @@ HUDLobby::refresh()
 {
 	HUD::refresh();
 	redSoulsText->setMessage(std::to_string(redSouls));
-	redSoulsText->setPos(relativeX(1870), relativeY(150 + getNumberY(redSouls)));
-	redSoulsText->setSize(getNumberSize(redSouls));
+	redSoulsText->setPos(relativeX((float)1870), relativeY((float)150 + getNumberY(redSouls)));
+	redSoulsText->setSize((int)getNumberSize(redSouls));
 	redSoulsText->render();
 }
 
 HUDBet::HUDBet(GameState* gS) : HUD(gS)
 {
-	betDescText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX(1600), relativeY(150),
-		relativeX(50), relativeX(3), Text::DERECHA);
+	betDescText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX((float)1600), relativeY((float)150),
+		relativeX((float)50), relativeX((float)3), Text::DERECHA);
 	betDescText->setMessage("APUESTA");
 	gS->addObjectsUI(betDescText);
-	betText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX(1870), relativeY(150 + getNumberY(bet)),
-		relativeX(getNumberSize(bet)), relativeX(3), Text::DERECHA);
+	betText = new Text(gS, gS->getGame()->getTypo(GRAND_CASINO), relativeX((float)1870), relativeY((float)150 + getNumberY(bet)),
+		relativeX((float)getNumberSize(bet)), relativeX((float)3), Text::DERECHA);
 	betText->setMessage(std::to_string(bet));
 	gS->addObjectsUI(betText);
 
@@ -103,8 +103,8 @@ HUDBet::refresh()
 {
 	HUD::refresh();
 	betText->setMessage(std::to_string(bet));
-	betText->setPos(relativeX(1870), relativeY(150 + getNumberY(bet)));
-	betText->setSize(getNumberSize(bet));
+	betText->setPos(relativeX((float)1870), relativeY((float)150 + getNumberY(bet)));
+	betText->setSize((int)getNumberSize(bet));
 	betText->render();
 }
