+++ DeathTokenBase/DeathToken/Slots.h
@@ -15,5 +15,5 @@ public:
 	inline Slots(Game* g) : GameState(g), ui(new UISlots(this, g, this)), hud(new HUDBet(this)), bet(0)
 		{ HUDManager::getHudBet()->refresh(); }
 	virtual void iniciarGiro() = 0;
-	inline void setBetTurno(int n) { bet = n; }
+	inline void setBetTurno(long long n) { bet = n; }
 };
